pyproject.toml
[project]
name = "slm-multi-agent-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets<4.0.0",
    "ddgs>=9.6.1",
    "docling>=2.61.2",
    "duckduckgo-search>=8.1.1",
    "e2b-code-interpreter>=2.3.0",
    "ipykernel>=7.0.1",
    "langchain>=1.0.5",
    "langchain-community>=0.4",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.1",
    "langgraph-api>=0.4.46",
    "ollama>=0.6.0",
    "openai>=2.7.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "playwright>=1.48.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.12",
]

[dependency-groups]
dev = [
    "langgraph-cli>=0.4.4",
    "pytest>=8.0.0",
]
